Refactor reference detail routing
