+++ client/src/js/indexes/components/__tests__/Indexes.js
@@ -1,6 +1,6 @@
-import { IndexesList } from "../List";
+import { Indexes } from "../Indexes";
 
-describe("<IndexesList />", () => {
+describe("<Indexes />", () => {
     let props;
 
     beforeEach(() => {
@@ -12,24 +12,24 @@ describe("<IndexesList />", () => {
     });
 
     it("should render", () => {
-        const wrapper = shallow(<IndexesList {...props} />);
+        const wrapper = shallow(<Indexes {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
 
     it("should render with empty documents array", () => {
         props.documents = [];
-        const wrapper = shallow(<IndexesList {...props} />);
+        const wrapper = shallow(<Indexes {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
 
     it("should render with [documents=null]", () => {
         props.documents = null;
-        const wrapper = shallow(<IndexesList {...props} />);
+        const wrapper = shallow(<Indexes {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
 
     it("should call onLoadNextPage on mount", () => {
-        shallow(<IndexesList {...props} />);
+        shallow(<Indexes {...props} />);
         expect(props.onLoadNextPage).toHaveBeenCalledWith(props.refId, 1);
     });
 });
