--- client/src/js/indexes/components/__tests__/Indexes.test.js
@@ -1,8 +1,0 @@
-import Indexes from "../Indexes";
-
-describe("<Indexes />", () => {
-    it("should render", () => {
-        const wrapper = shallow(<Indexes />);
-        expect(wrapper).toMatchSnapshot();
-    });
-});
