+++ client/src/js/otus/components/List.js
@@ -3,6 +3,8 @@ import React from "react";
 import { connect } from "react-redux";
 import { LoadingPlaceholder, NoneFound, ScrollList } from "../../base";
 import RebuildAlert from "../../indexes/components/RebuildAlert";
+import ReferenceDetailHeader from "../../references/components/Detail/Header";
+import ReferenceDetailTabs from "../../references/components/Detail/Tabs";
 import { findOTUs } from "../actions";
 import { getTerm } from "../selectors";
 import CreateOTU from "./Create";
@@ -29,6 +31,9 @@ class OTUsList extends React.Component {
 
         return (
             <div>
+                <ReferenceDetailHeader />
+                <ReferenceDetailTabs />
+
                 <RebuildAlert />
                 <OTUToolbar />
                 <CreateOTU {...this.props} />
