+++ client/src/js/references/components/Detail/Detail.js
@@ -2,7 +2,7 @@ import { get } from "lodash-es";
 import React from "react";
 import { connect } from "react-redux";
 import { Redirect, Route, Switch } from "react-router-dom";
-import { Badge, LoadingPlaceholder, NotFound, TabLink, Tabs } from "../../../base";
+import { LoadingPlaceholder, NotFound } from "../../../base";
 import IndexDetail from "../../../indexes/components/Detail";
 import Indexes from "../../../indexes/components/Indexes";
 import OTUDetail from "../../../otus/components/Detail/Detail";
@@ -10,9 +10,8 @@ import OTUList from "../../../otus/components/List";
 import { checkRefRight } from "../../../utils/utils";
 import { getReference } from "../../actions";
 import EditReference from "./Edit";
-import ReferenceDetailHeader from "./Header";
 import ReferenceManage from "./Manage";
-import { ReferenceSettings } from "./Settings";
+import ReferenceSettings from "./Settings";
 
 class ReferenceDetail extends React.Component {
     constructor(props) {
@@ -25,58 +24,33 @@ class ReferenceDetail extends React.Component {
             return <NotFound />;
         }
 
-        if (this.props.detail === null || this.props.detail.id !== this.props.match.params.refId) {
+        if (!this.props.id || this.props.id !== this.props.match.params.refId) {
             return <LoadingPlaceholder />;
         }
-
-        const { id, otu_count, remotes_from } = this.props.detail;
 
         return (
             <div className="detail-container">
                 <Switch>
                     <Route path="/refs/:refId/otus/:otuId" component={OTUDetail} />
                     <Route path="/refs/:refId/indexes/:indexId" component={IndexDetail} />
-                    <Route
-                        path="/refs"
-                        render={() => (
-                            <div>
-                                <ReferenceDetailHeader />
-
-                                <Tabs>
-                                    <TabLink to={`/refs/${id}/manage`}>Manage</TabLink>
-                                    <TabLink to={`/refs/${id}/otus`}>
-                                        OTUs <Badge>{otu_count}</Badge>
-                                    </TabLink>
-                                    <TabLink to={`/refs/${id}/indexes`}>Indexes</TabLink>
-                                    <TabLink to={`/refs/${id}/settings`}>Settings</TabLink>
-                                </Tabs>
-
-                                <Switch>
-                                    <Redirect from="/refs/:refId" to={`/refs/${id}/manage`} exact />
-                                    <Route path="/refs/:refId/manage" component={ReferenceManage} />
-                                    <Route path="/refs/:refId/otus" component={OTUList} />
-                                    <Route path="/refs/:refId/indexes" component={Indexes} />
-                                    <Route
-                                        path="/refs/:refId/settings"
-                                        render={() => <ReferenceSettings isRemote={remotes_from} />}
-                                    />
-                                </Switch>
-
-                                <EditReference />
-                            </div>
-                        )}
-                    />
+                    <Redirect from="/refs/:refId" to={`/refs/${this.props.id}/manage`} exact />
+                    <Route path="/refs/:refId/manage" component={ReferenceManage} />
+                    <Route path="/refs/:refId/otus" component={OTUList} />
+                    <Route path="/refs/:refId/indexes" component={Indexes} />
+                    <Route path="/refs/:refId/settings" component={ReferenceSettings} />
                 </Switch>
+
+                <EditReference />
             </div>
         );
     };
 }
 
 const mapStateToProps = state => ({
+    canModify: checkRefRight(state, "modify"),
     error: get(state, "errors.GET_REFERENCE_ERROR", null),
-    detail: state.references.detail,
-    pathname: state.router.location.pathname,
-    canModify: checkRefRight(state, "modify")
+    id: get(state, "references.detail.id"),
+    pathname: state.router.location.pathname
 });
 
 const mapDispatchToProps = dispatch => ({
